Add Canadian Cities to generation script
