+++ utils/getCities.js
@@ -1,0 +1,49 @@
+var request = require('request');
+var fs = require('fs');
+var unzipper = require('unzipper');
+
+function getDataForCountry(country_code){
+  let country_code_lower = country_code.toLowerCase();
+  let country_code_upper = country_code.toUpperCase();
+  var states = JSON.parse(fs.readFileSync("../data/"+country_code_lower+"-states.json"));
+
+  unzipper.Open.url(request,'http://download.geonames.org/export/zip/'+country_code_upper+'.zip')
+    .then(function(d) {
+      var file = d.files.filter(function(d) {
+        return d.path === country_code_upper+'.txt';
+      })[0];
+      return file.buffer();
+    })
+    .then(function(d) {
+      var cityHash = {};
+      for (var key in states) {
+        cityHash[states[key]] = [];
+      }
+      console.log(cityHash);
+      var citiesData = d.toString().split("\n");
+      citiesData.forEach(function(element) {
+        var cityData = element.split('\t');
+        //console.log(cityData[2] + ", " + cityData[4]);
+        if (cityData[4] && cityData[4].length == 2 && cityHash.hasOwnProperty(cityData[4])) {
+          cityHash[cityData[4]].push(cityData[2]);
+        }
+      });
+      //Remove duplicates and sort from longest to shortest
+      for (var key in cityHash) {
+        var uSet = new Set(cityHash[key]);
+        cityHash[key]= [...uSet];
+        cityHash[key].sort(function(a, b) {
+          return b.length - a.length;
+        });
+      }
+      fs.writeFile('../data/'+country_code_lower+'-cities.json', JSON.stringify(cityHash), 'utf8', function() {
+        console.log(cityHash);
+      });
+
+    });
+
+
+}
+
+getDataForCountry("CA");
+getDataForCountry("US");
