+++ utils/package.json
@@ -1,5 +1,6 @@
 {
   "dependencies": {
+    "package.json": "^2.0.1",
     "request": "*",
     "unzipper": "*"
   }
